+++ cypress.config.js
@@ -7,6 +7,7 @@ module.exports = defineConfig({
     setupNodeEvents(on, config) {
       return require("./cypress/plugins/index.js")(on, config);
     },
-    excludeSpecPattern: process.env.CI ? ["cypress/e2e/all.cy.js"] : []
+    excludeSpecPattern: process.env.CI ? ["cypress/e2e/all.cy.js"] : [],
+    video: false
   }
 });