+++ changelog.md
@@ -1,5 +1,11 @@
 # goNORTH Changelog
 
+## 0.4.1
+
+* 2022-10-11 Disabled Cypress video.
+* 2022-10-10 Allowed items to be taken when they have no container.
+* 2022-10-08 Allowed containers to be given keys by name.
+
 # 0.4.0
 
 * 2022-10-07 Made containers unlockable with or without keys.